added more columns in the user_stats, added more options in the user_goal page and able to fill the user_stats table

+++ src/db/migrations/20230429213132_create_user_stats.js
@@ -6,8 +6,8 @@ exports.up = (knex) => {
     return knex.schema.createTable('user_stats', (table) => {
         table.integer('user_id').notNullable();
         table.foreign("user_id").references("id").inTable("users");
-        table.integer('height').notNullable();
-        table.integer('weight').notNullable();
+        table.float('height').notNullable();
+        table.float('weight').notNullable();
         table.float('bmi').notNullable();
         table.timestamp('created_at').defaultTo(knex.fn.now());
     })
